Set up saving of model objects in Firebase Database
By default, reading and writing the Firebase database is only
allowed for authenticated users (real user or anonymous). Had to
set read and write access to "true" in the Firebase console website
of the project to allow writing to the database at this stage.

Next step is to set up an authenticated anonymous user in order
to restore the database read and write rules to "auth != null".