Enable signing in to Google in order to auth to Firebase
This requires to enable the "Google" sign-in method of the Firebase
console website for the project. Other available Firebase sign-in
methods are "Email/Password" and "Anonymous".

The user is redirected to the SignInActivity whenever the app is
started and there is no authenticated Firebase user. On the
SignInActivity, the user can sign in to their Google account.
This account is then used to sign in to Firebase.

Being signed-in to Firebase allows to restore the database read
and write access to "auth != null".