createDeal and getAllDeals method of the database API
createDeal: allow to save a new deal in the database (the auto-
generated key of the Firebase database is taken as the ID of the deal).

getAllDeals: get an ArrayList of all the deals that are currently
in the database.

Currently, the MainActivity has a normal RecyclerView and adapter
and is notified by the DatabaseManager through a listener callback
method when a new deal was added. This adapte could be replaced
with a FirebaseRecyclerAdapter, which automatically synchronises
with the database, i.e. no manual updating of the adapter is
necessary.