Database interface "create" and "get" methods
Create: save a Deal, Customer, or Business in the database by including
an auto-generated ID. The ID is returned to the caller.

Get: get a Deal, Customer, or Business object from the database by
supplying its ID. The result object is returned in a QueryCallbackSingle
(method "yourResult") which must be passed to the "get" method.