Denormalising basic data for Firebase Database
The one-to-many relationships (Deal/Business) and many-to-many
relationships (Customer/Deal represented by Acquisition) are saved
in a denormalised way in the Firebase Database. This means:

- A business holds a list of the IDs of all the deals it publishes
- A customer holds a list of all its acquisitions
- A deal holds a list of all its acquisitions

This requires the insertion of the above redundant information
whenever a new deal or acquisitions, respectively, is created.

This approach is officially recommended to make certain queries
(e.g. get all deals published by a certain business) faster
(https://firebase.googleblog.com/2013/04/denormalizing-your-data-is-normal.html)

At this point, the following queries of the DatabaseManager are
fully functional:

- create(Customer|Business|Deal|Acquisition)
- get(Customer|Business|Deal|Acquisition)