Exploring the Firebase Database query API
This commit contains code in DatabasManager.java that will be deleted
in the next commit, but could be useful for future use cases (use a
single value event listener to detect when all the initial data has
been loaded by a child event listener; this can be used e.g. to
handle only the new incremental data with the child event listener
and ignore the existing inital data).