More exploring of the Firebase Database API
For retrieving an displaying lists of data, no query methods are needed,
but the used RecyclerView is backed by a FirebaseRecyclerAdapter
in which we deposit a Firebase "Query". The adapter then automatically
updates the content of the list based on the constraints of the query
and the content of the database. This is because a Firebase "Query" is
not a one-shot query, but it is always active because it is actually a
set of listener callbacks.

The query placed in the FirebaseRecyclerAdapter may contain order by
(orderByChild, orderByKey, orderByValue), limit (limitToFirst,
limitToLast), and range (startAt, endAt, equalTo) methods.

The implementation of the FirebaseRecyclerAdapter is quite simple
and it could be reimplemented and customised.

For retrieving individual objects by ID, it's ok to emulate one-shot
queries in the corresponding methods of DatabaseManager.