Custom queries for FirebaseRecyclerAdapter
The FirebaseRecyclerAdapter (used by each fragment of the ViewPager
in the HomeActivity) is passed a custom Query. Currently, the queries
for the three fragments are "order by deal price", "order by regular
price", and "order by title", each one limited to the "first 5 items".

The content of the lists is updated automatically by the Firebase-
RecyclerAdapter based on the query and the content of the database.
If for example a new deal with a lowest deal price is created, it
is automatically displayed in the list and another deal (which is
now not anymore in the top 5 lowest deal prices) is removed from
the list.

This is the way to display lists of data with Firebase Database. One
problem that needs to be solved is pagination, wich is not directly
suppoted by the Firebase API, but it can be achieved with some
effort (https://github.com/firebase/FirebaseUI-Android/pull/26).