Create FirebaseRecyclerAdapter to be used with denormalised data
The relations "deals of business", "acquisitions of customer", and
"acquisitions of deal" are denormalised in the database, i.e.
Business contains a list of deal IDs, Customer contains a list of
acquisition IDs, and Deal contains a list of acquisition IDs. This
is to make these lookups more efficient and scalable.

However, this requires a modified FirebaseRecyclerAdapter, because
the adapter now listens on the list of IDs and not on the objects
themselves. So in the populateViewHolder method we need to retrieve
the object itself by adding another listener to its database location.