Generify DenormFirebaseRecyclerAdapter
This is a FirebaseRecyclerAdapter listening to a set of data objects
that is pointed to by a set of IDs in another object (i.e. the
relations between these objects are denormalised to improve lookup
performance). The adapter listens to the IDs, but when creating
an item view, fetches the real object belonging to the ID and
populates the list item view with data of this object.