Remove Acquisition model and denormalise its data
The data of an acquisition (acquisition date and redemption date) is
now saved under the deals list of the corresponding customer
(customer/deals/dealId) and under the customers list of the corre-
sponding deal (deal/customers/customerId).

In customer/deals/dealId additionaly the begin and end validity date
is saved. This allows to efficiently start a query on deals of a
customer which are upcoming, current, or expired.