Add sample activity querying and displaying a single customer
Directly add a value event listener to the customer refererenc in
the activity. The 'getCustomer' method of DatabaseManager is not
used, because the added listener must be removed when the activity
is disposed of.

Possible improvement: create an abstract activity taking care of
adding and removing the listener in onStart and onStop.