Finish basic version of sign-up/sign-in
Works for the most part.

The Google sign in options needs to be improved: clicks on the
"Sign in with Google" button should be handled in a uniform way.
When getting the Google sign in result (onActivityResult) we should
check whether a Firebase user with the corresponding email address
already exists. If yes, there's nothing to do, if no, we need to
prompt the user whether she wants to be a customer or business,
sign in to Firebase, and then save the customer or business in the
database.

Note: there can be no two Firebase users with the same email address.
If signing up "with email and password" with a Google account email
address, and later "signing in with Google" with the account belonging
to this same email address, then the user is signed in to the initially
created account.