Add CustomerDealsActivity
...with two tabs "Open" and "Closed":

- Open: deals whose validity end time is not in the past AND which
  have not been redeemed
- Closed: deals which have been redeemed OR whose validity end time
  is in the past

All deals are sorted by the validity end time: Open -> smallest first,
Closed -> largest first.

Note: the current solution is very ugly: get all the deals and then
in the FirebaseRecyclerAdapter (populateViewHolde()) for each
RecyclerView, hide the items which should not be displayed by setting
their height to 0. The reason that this cannot be easily solved better
is that with Firebase it's not possible ot sort and filter on
different or multiple fields.